Initial commit of all decompiled code. A huge note about this commit is that the code, due to decompiler idiosyncracies and programming by LordFFA, is entirely broken and not in working order.
Another HUGE note is that these files have all been reformatted using Eclipse's autoformat algorithm (so extraneous spacing that may have been in the original source files is now gone, it was an eyesore).

The primary purpose of this commit is to serve as a baseline for what was modified.